Move pagination functionality to separate module
