+++ flask/boxwise_flask/graph_ql/resolvers.py
@@ -1,6 +1,4 @@
 """GraphQL resolver functionality"""
-import base64
-
 from ariadne import MutationType, ObjectType, QueryType, convert_kwargs_to_snake_case
 from peewee import fn
 
@@ -26,6 +24,7 @@ from ..models.size import Size
 from ..models.transaction import Transaction
 from ..models.user import User
 from ..models.x_beneficiary_language import XBeneficiaryLanguage
+from .pagination import decode_cursor, generate_page, pagination_parameters
 
 query = QueryType()
 mutation = MutationType()
@@ -38,58 +37,6 @@ product = ObjectType("Product")
 product_category = ObjectType("ProductCategory")
 qr_code = ObjectType("QrCode")
 user = ObjectType("User")
-
-
-class PageInfo:
-    """Container for pagination information."""
-
-    def __init__(self, *, has_next_page=False, start_cursor=None):
-        self.has_next_page = has_next_page
-        self.start_cursor = start_cursor
-
-
-def pagination_parameters(pagination_input):
-    """Retrieve cursor and limit (default: None and 50, resp.) from the given pagination
-    input dictionary.
-    """
-    limit = 50
-    if pagination_input is None:
-        return None, limit
-    return pagination_input.get("cursor"), pagination_input.get("limit", limit)
-
-
-def decode_cursor(model, cursor):
-    """Decode given cursor (a base64-encoded string) into a condition that can be
-    plugged into a ModelSelect.where() clause for the given model.
-    The cursor serves as a start value for the query (default: 1).
-    """
-    start_value = 1 if cursor is None else int(base64.b64decode(cursor))
-    return model.id >= start_value
-
-
-def generate_page_info(*, elements, limit):
-    """Generate pagination information from given elements and page limit. The elements
-    comprise the current page and possibly the first element of the next page.
-    If the size of the elements exceeds the limits, a next page exists.
-    The cursor for the next page is derived from the ID of the last element: the ID is
-    zero-padded to a byte-string of length 8, and then base64-encoded.
-    """
-    info = PageInfo()
-    if len(elements) > limit:
-        info.has_next_page = True
-        info.start_cursor = base64.b64encode(f"{elements[-1].id:08}".encode()).decode()
-    return info
-
-
-def generate_page(*, elements, limit):
-    """Return a GraphQL Page type wrapping the given elements, and including appropriate
-    page info.
-    """
-    page_info = generate_page_info(elements=elements, limit=limit)
-    return {
-        "elements": elements[:-1] if page_info.has_next_page else elements,
-        "page_info": page_info,
-    }
 
 
 @user.field("bases")
